Add use-effect, use-state impls
This adds working hooks to tui, which can be used anywhere in render.

Hooks must be used unconditionally, they behave the same way as in
vanilla react: you are given a 'slot' and may use that slot to keep
state across renders. tui-hooks comes with 'reference' types used in
the implementation of use-effect, use-state.