Reverse dependencies, effect
This is a step towards "automatic" macros: emacs lisp can't really
compare closures correctly and dependencies that aren't lined up with
the closure result in confusing bugs, so we might want to do it
ourselves.

This invariably will be pretty complex but ergonomics of it would be
kind of fantastic: typescript has a linter but the emacs-lisp macro
system could actually give us the 'pre-process step' of 'analyze this
form and see what it depends on'.